fix: refactoring - separated operations db and logics, delete conditionals unnecesary
